dmd: Check unit tests too
Pass -unittest to DMD in the D checker by default.

This is useful when writing template code, which is only instantiated in unittest blocks. It also allows making sure the unittest blocks are semantically valid, without needing to actually run them.

Does not not actually run the unittests, only verifies the correctness of their code.